minor refactor Scaffolding of backend fix #1
