+++ server/src/main/java/edu/gmu/server/controller/PlayController.java
@@ -1,0 +1,11 @@
+package edu.gmu.server.controller;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@Slf4j
+@RestController
+@RequestMapping("/api/play")
+public class PlayController {
+}
