+++ server/src/main/java/edu/gmu/server/service/AuthService.java
@@ -4,38 +4,22 @@ import edu.gmu.server.dto.AuthenticateDto;
 import edu.gmu.server.dto.RegisterDto;
 import edu.gmu.server.entity.Stats;
 import edu.gmu.server.entity.User;
-import edu.gmu.server.exception.HeartsBadCredentialsException;
-import edu.gmu.server.exception.HeartsResourceNotFoundException;
 import edu.gmu.server.repository.StatsRepository;
 import edu.gmu.server.repository.UserRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.http.HttpStatus;
 import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.server.ResponseStatusException;
 
-import javax.sql.rowset.serial.SerialBlob;
 import javax.transaction.Transactional;
-import java.io.IOException;
-import java.sql.Blob;
-import java.sql.SQLException;
 import java.time.ZoneOffset;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
-import java.util.Map;
 import java.util.Optional;
 
 @Slf4j
@@ -54,11 +38,6 @@ public class AuthService {
     this.authenticationManager = authenticationManager;
   }
 
-  @Transactional
-  public Page<User> getAllUsers(int page, int limit, Map<String, String> filters) {
-    Pageable pageable = PageRequest.of(page, limit);
-    return this.userRepository.findAll(pageable);
-  }
 
   @Transactional
   public Optional<User> register(RegisterDto registerDto) {
@@ -94,43 +73,6 @@ public class AuthService {
   }
 
   @Transactional
-  public void uploadProfilePicture(UserDetails principal, MultipartFile picture) throws IOException, SQLException {
-    if (principal != null) {
-      log.info("principal {} attempts to upload a profile picture", principal.getUsername());
-      User currentUser = this.userRepository.findByUsername(principal.getUsername())
-        .orElseThrow(() -> new ResponseStatusException(HttpStatus.UNAUTHORIZED, ""));
-      log.info("Picture info: {}, {}, {}, {}", picture.getName(), picture.getSize(),
-        picture.getBytes().length, picture.getOriginalFilename());
-      Blob picBlob = new SerialBlob(picture.getBytes());
-      currentUser.setProfilePicture(picBlob);
-    } else {
-      log.info("Attempt access to profile picture with bad credentials");
-      throw new HeartsBadCredentialsException("Bad credentials provided");
-    }
-  }
-
-  public byte[] getProfilePicture(UserDetails principal)
-    throws HeartsBadCredentialsException, HeartsResourceNotFoundException {
-    try {
-      if (principal != null) {
-        log.info("principal {} attempts to get profile picture", principal.getUsername());
-        User currentUser = this.userRepository.findByUsername(principal.getUsername())
-          .orElseThrow(() -> new ResponseStatusException(HttpStatus.FORBIDDEN, "Invalid Credentials"));
-        int len = (int) currentUser.getProfilePicture().length();
-        return currentUser.getProfilePicture().getBytes(1, len);
-      } else {
-        throw new BadCredentialsException("Bad Credentials provided");
-      }
-    } catch (BadCredentialsException e) {
-      log.info("Attempt access to profile picture with bad credentials");
-      throw new HeartsBadCredentialsException("Bad Credentials provided");
-    } catch (NullPointerException e) {
-      throw new HeartsResourceNotFoundException("A profile picture not found");
-    } catch (Exception e) {
-      throw new HeartsResourceNotFoundException("Operation failed");
-    }
-  }
-
   public Optional<User> authenticate(AuthenticateDto authenticateDto) {
     return this.loadAuthenticatedUser(authenticateDto.getUsername(), authenticateDto.getPassword());
   }
