+++ server/src/main/java/edu/gmu/server/service/PlayService.java
@@ -1,0 +1,4 @@
+package edu.gmu.server.service;
+
+public class PlayService {
+}
